refactor(libs) reference lookup eslint requirements
This commit refactors the reference lookup library according to
eslint standards. It also renames the library fixing a language
typo.
